Make authenticated middleware written in Typescript
Later, we have to use dynamic imports in order to enable type checking.
However dynamic imports will somehow lead to build failure of client-side.
I assume it's because in client-side NuxtJS tries to import the libraries that
should be imported only in server-side. Though I have no idea of avoiding this.